Move Most Viewed Pages Macro to livetable #10 * I modified the dependencies and removed some parameters.
