+++ application-analytics-default/src/main/java/com/xwiki/analytics/script/AnalyticsScriptService.java
@@ -31,7 +31,6 @@ import org.xwiki.stability.Unstable;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.xwiki.analytics.AnalyticsManager;
-import com.xwiki.analytics.configuration.AnalyticsConfiguration;
 
 /**
  * The scripting service for the Analytics App (Pro).
@@ -45,26 +44,22 @@ import com.xwiki.analytics.configuration.AnalyticsConfiguration;
 @Singleton
 public class AnalyticsScriptService implements ScriptService
 {
-    @Inject
-    private AnalyticsConfiguration configuration;
 
     @Inject
     @Named("Matomo")
     private AnalyticsManager analyticsManager;
-
     /**
      * This Method is used to interrogate the Matomo API.
      *
-     * @param jsonNormaliserHint hint to select the json normaliser.
-     * @param parameterList a map of the parameters for the url
+     * @param jsonNormaliserHint A hint to select the JSON normalizer. This hint is needed because Matomo returns
+     * JSON in various formats. With this hint, I can switch the normalizer at runtime.
+     * @param parameters a map of the parameters for the url
      * @return will return a json string.
      */
-    public JsonNode getDataFromRequest(Map<String, String> parameterList, String jsonNormaliserHint)
+    public JsonNode getDataFromRequest(Map<String, String> parameters, String jsonNormaliserHint)
     {
-        parameterList.put("idSite", configuration.getIdSite());
-        parameterList.put("token_auth", configuration.getAuthenticationToken());
         try {
-            return analyticsManager.requestData(configuration.getRequestAddress(), parameterList, jsonNormaliserHint);
+            return analyticsManager.requestData(parameters, jsonNormaliserHint);
         } catch (Exception e) {
             throw new RuntimeException(e);
         }
